improved componentization and implemented the mobile version for the game
