+++ src/components/Game/styles.ts
@@ -1,8 +1,18 @@
-import styled from "styled-components";
+import { ImSpinner9 } from "react-icons/im";
+import styled, { keyframes } from "styled-components";
 
 export const Game = styled.main`
   flex: 1;
   position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  @media (max-width: 860px) {
+    justify-content: flex-start;
+    gap: 32px;
+  }
 `;
 
 export const Header = styled.h1`
@@ -12,71 +22,40 @@ export const Header = styled.h1`
   font-family: Arial, Helvetica, sans-serif;
 `;
 
-export const CountdownContainer = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: 2;
-`;
-
-export const Board = styled.div`
-  background-color: #202020;
-  width: 500px;
-  height: 500px;
-  border-radius: 50%;
-  padding: 64px;
-
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: 1fr 1fr;
-  grid-auto-rows: 150px;
-  grid-auto-columns: 150px;
-  gap: 64px;
-
-  box-shadow: 0 0 4px 4px rgba(0, 0, 0, 0.25);
-  border: 5px solid #c77225;
-`;
-
-export const BoardButton = styled.button<{ color: string; active: boolean }>`
-  width: 100%;
-  height: 100%;
-  border-radius: 64px;
-  background: ${({ color }) => color};
-  border: none;
-  transition: 600ms;
-
-  filter: ${({ active }) => `brightness(${active ? "3" : "1"})`};
-
-  &:hover {
-    cursor: pointer;
+export const MobileMenu = styled.div`
+  @media (min-width: 861px) {
+    display: none;
   }
 `;
 
-export const BoardCenter = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background: #202020;
-
-  width: 250px;
-  height: 250px;
-  border-radius: 50%;
-
-  box-shadow: 0 0 4px 4px rgba(0, 0, 0, 0.25);
+export const Icon = styled.span`
+  font-size: 48px;
+  color: #c77225;
+  text-align: center;
+  margin-top: 40px;
+  transition: 300ms;
 
   display: flex;
   align-items: center;
   justify-content: center;
+
+  &:hover {
+    color: white;
+    cursor: pointer;
+  }
 `;
 
-export const Image = styled.img`
-  width: 150px;
-  height: 150px;
+const spin = keyframes`
+  from{
+    transform-origin: center;
+    transform: rotate(0);
+  } to{
+    transform-origin: center;
+    transform: rotate(360deg);
+  }
+`;
+
+export const Spinner = styled(ImSpinner9)`
+  transform-origin: center;
+  animation: ${spin} 1200ms linear infinite;
 `;
